+++ bfabric_app_runner/src/bfabric_app_runner/dispatch/dispatch_resource_flow.py
@@ -7,6 +7,7 @@ import polars as pl
 import yaml
 from bfabric import Bfabric
 from bfabric.entities import Resource
+from bfabric.entities.core.reader_utils import present_entities
 from bfabric.experimental.workunit_definition import WorkunitDefinition
 from bfabric.utils.cli_integration import use_client
 from pandera import Field
@@ -86,11 +87,7 @@ class ResourceDispatcher:
     @staticmethod
     def _build_input_resources_df(resource_ids: list[int], client: Bfabric) -> DataFrame[InputTable]:
         """Creates the InputTable DataFrame from the list of resource IDs."""
-        resources = [
-            resource
-            for resource in client.reader.read_ids("resource", resource_ids, expected_type=Resource).values()
-            if resource is not None
-        ]
+        resources = present_entities(client.reader.read_ids("resource", resource_ids, expected_type=Resource))
         if not resources:
             raise ValueError("No resources to dispatch")
         attributes = [